requirements.txt
pip==21.0.1
setuptools==54.2.0
selenium==3.141.0
pytest==6.2.2
webdriver-manager==3.3.0
python-dotenv==0.16.0